requirements.txt
Django~=4.2.1
numpy~=2.1.1
matplotlib~=3.9.2
djangorestframework~=3.15.2